sleep alerting system when eyes are closed
A brief description of what this project does and who it's for

On average 1200 rest gets effected badly. The major reason of these accidents is drowsiness caused by both sleep and alcohol. Due to driving for long time or intoxication, drivers might feel sleepy which is the biggest distraction for them while driving. This distraction might cost death of driver and other passengers in the vehicle and at the same time it also causes death of people in the other vehicles and pedestrians too. This mistake of one person on road would take their own life and also takes lives of other and put respective families in sorrow and tough situations. To prevent such accidents we, team 5A propose a system which alerts the driver if he/she feels drowsy. To accomplish this, we implement the solution using computer-vision based machine learning model. The driver’s face is detected by face recognition algorithm continuously using a camera and the face of the driver is captured. The face of the driver is given as input to a classification algorithm which is trained with a data set of images of drowsy and non-drowsy faces. The algorithm uses landmark detection to classify the face as drowsy or not drowsy.